# droneport-server/src/main/java/com/hitachi/droneroute/dpm/dto/DronePortReserveInfoListResponseDto.java
package com.hitachi.droneroute.dpm.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonInclude.Include;
import java.util.List;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

/** 離着陸場予約情報一覧取得応答DTO */
@NoArgsConstructor
@Setter
@Getter
@ToString
public class DronePortReserveInfoListResponseDto {

  /** 離着陸場予約情報リスト */
  private List<DronePortReserveInfoListElement> data;

  /** 1ページ当たりの件数 */
  @JsonInclude(Include.NON_NULL)
  private Integer perPage;

  /** 現在ページ番号 */
  @JsonInclude(Include.NON_NULL)
  private Integer currentPage;

  /** 最終ページ番号 */
  @JsonInclude(Include.NON_NULL)
  private Integer lastPage;

  /** 全体件数 */
  @JsonInclude(Include.NON_NULL)
  private Integer total;
}